+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## [Unreleased]
 
+### Improvement
+- Improve the informativeness of error outputs from the token validations
+
 ## [0.0.43]
 ### Changed
 - Breaking change to the json schema, in order to use jackson